backend/requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Environment and Configuration
python-dotenv>=1.0.0
python-decouple>=3.8

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10

# HTTP and API
httpx>=0.28.0
aiohttp>=3.11.0

# HTML/XML Parsing
beautifulsoup4>=4.12.3
lxml>=5.3.0
html5lib>=1.1

# Retry Logic and Resilience
tenacity>=9.0.0

# CORS and Middleware
python-multipart>=0.0.20

# Date/Time Handling
python-dateutil>=2.9.0

# Text Processing and NLP (Optional)
nltk>=3.9.1
textblob>=0.18.0

# Data Processing - Using newer versions compatible with Python 3.13
pandas>=2.2.0
numpy>=2.0.0

# Web Scraping and RSS Parsing
feedparser>=6.0.11

# AI/ML Libraries for Embeddings and Vector Search
sentence-transformers>=3.3.0
transformers>=4.47.0
torch>=2.5.0

# Vector Database
chromadb>=0.5.23

# Development and Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.10.0
isort>=5.13.0
flake8>=7.1.0
mypy>=1.13.0

# Logging and Monitoring
structlog>=24.4.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.1
